refactor(permanents): Refactor the collective submatrices permanents calculators by moving most of the common code to the base class.
